+++ prijateli_tree/app/routers/games.py
@@ -17,7 +17,13 @@ from prijateli_tree.app.database import (
     GameSessionPlayer,
     get_db,
 )
-from prijateli_tree.app.routers.game_utils.utils import (
+from prijateli_tree.app.utils.constants import (
+    DENIR_FACTOR,
+    FILE_MODE_READ,
+    STANDARD_ENCODING,
+    WINNING_SCORE,
+)
+from prijateli_tree.app.utils.games import (
     did_player_win,
     get_bag_color,
     get_current_round,
@@ -26,12 +32,6 @@ from prijateli_tree.app.routers.game_utils.utils import (
     get_previous_answers,
     get_session_player_from_player,
     raise_exception_if_none,
-)
-from prijateli_tree.app.utils.constants import (
-    DENIR_FACTOR,
-    FILE_MODE_READ,
-    STANDARD_ENCODING,
-    WINNING_SCORE,
 )
